fix: preserve pipeline payloads for call logs
